+++ app/controllers/site_page_controller.rb
@@ -95,6 +95,8 @@ class SitePageController < ApplicationController
       gon.analysis_widgets = @setting.widgets.blank? ? nil : (JSON.parse @setting.widgets)
       # ... data
       gon.analysis_data = @setting.get_filtered_dataset
+      # ... metadata
+      gon.metadata = Dataset.get_metadata_for_frontend(session[:user_token], @setting.dataset_id)
       # ... last modification of the fields
       gon.analysis_timestamp = @setting.fields_last_modified
       # ... legend fields