+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 - Fix issue on dataset/widget filter step.
 - Adds support for more datatypes in chart widgets.
+- Added metadata table modal to management section.
+- Added metadata table modal to open content widgets.
+- Added metadata table modal to analysis dashboards.
+
+### 25 April 2017
 - Add metadata fields to dataset creation and list
 - Improved 404 page
 - Added sitemap.xml and sitemap.html pages